My first library used, learning how libaries

+++ app.py
@@ -3,7 +3,7 @@ PrecioApp = int(input("PrecioApp: "))
 Meses = int(input("Meses: "))
 
 Ganacia = Reseñas * PrecioApp
-GanaciaPorMes = int(Ganacia / Meses)
+GanaciaPorMes = Ganacia/Meses
 
 print("Ganacia total:", Ganacia , "MXN")
 print("Ganacia por mes:", GanaciaPorMes, "MXN" )
